prototipados y navegación de cosas varias

+++ frontend-android/PrototipadoComunidades/proyectobase2425/app/src/main/java/com/pmdm/proyectobase2425/models/Evento.kt
@@ -3,7 +3,7 @@ package com.pmdm.proyectobase2425.models
 import java.time.LocalDateTime
 
 data class Evento(
-    var eventoId: Int,
+    var eventoId: Long,
     var nombre: String,
     var ubicacion: String,
     var descripcion: String,
