+++ frontend-android/PrototipadoComunidades/proyectobase2425/app/src/main/java/com/pmdm/proyectobase2425/models/Producto.kt
@@ -1,7 +1,7 @@
 package com.pmdm.proyectobase2425.models
 
 data class Producto(
-    var productoId: Int,
+    var productoId: Long,
     var nombre: String,
     var descripcion: String,
     var puntos: Int
